+++ Danzer D3D12 Engine/Code/Engine/Rendering/Screen Rendering/LightHandler.cpp
@@ -11,7 +11,7 @@ LightHandler::~LightHandler(){}
 UINT LightHandler::CreatePointLight()
 {
     m_pointLightBuffer.emplace_back(PointLightBuffer());
-    UINT id = m_pointLightBuffer.size() - 1;
+    UINT id = (UINT)m_pointLightBuffer.size() - 1;
     m_pointLightBuffer[id].Init(m_framework.GetDevice(), &m_framework.CbvSrvHeap(), m_pointLightBuffer[id].FetchData(), sizeof(PointLightBuffer::Data));
-    return m_pointLightBuffer.size() - 1;
+    return (UINT)m_pointLightBuffer.size() - 1;
 }
